Fixed issue where sketch would send a double instead of an int and vi… (bp #189)

+++ lib/input/sketch/entities/layers/abstract_layer.dart
@@ -33,7 +33,7 @@ abstract class SketchNode implements DesignNode {
   final bool nameIsFixed;
   final dynamic resizingConstraint;
   final dynamic resizingType;
-  final int rotation;
+  final double rotation;
   final dynamic sharedStyleID;
   final bool shouldBreakMaskChain;
   final bool hasClippingMask;
